1. npc &플레이어 공격, exp, level, hp 연동 2. npc random 이동 구현 3. 채팅 구현 (but 보내는건 안됨 imgui 이슈) 4. npc 초기화 lua 연동 5. sector 구체화 --- but npc 죽었는데도 계속 살아움직이는 버그 발생 기말고사 보고 돌아온다.

+++ GSP_CLIENT/GSP_CLIENT/NPC.h
@@ -6,11 +6,11 @@ class CNPC
 private:
 	bool m_bActive;
 	char m_name[NAME_SIZE];
-	short m_id;
+	int m_id;
 	short m_x;
 	short m_y;
-	short m_level;
-	short m_hp;
+	int m_level;
+	int m_hp;
 
 	unsigned m_move_time;
 
@@ -64,17 +64,17 @@ public:
 	void SetTexture(sf::Texture& t, int left, int top) { m_sprite.setTexture(t); m_sprite.setTextureRect(sf::IntRect(left, top, TILE_SIZE, TILE_SIZE)); }
 	void SetActive(bool bTrue) { m_bActive = bTrue; }
 	void SetName(char* name) { strncpy_s(m_name, name, NAME_SIZE); }
-	void SetId(short id) { m_id = id; }
-	void SetHp(short hp) { m_hp = hp; }
+	void SetId(int id) { m_id = id; }
+	void SetHp(int hp) { m_hp = hp; }
 	void SetX(short x) { m_x = x; }
 	void SetY(short y) { m_y = y; }
 
 	char* GetName() { return m_name; }
-	short GetId() { return m_id; }
-	short GetX() { return m_x; }
-	short GetY() { return m_y; }
-	short GetLevel() { return m_level; }
-	short GetHp() { return m_hp; }
+	int GetId() { return m_id; }
+	int GetX() { return m_x; }
+	int GetY() { return m_y; }
+	int GetLevel() { return m_level; }
+	int GetHp() { return m_hp; }
 
 	void Move(short x, short y) { m_x = x; m_y = y; }
 	void Attack(short x, short y);
