+++ GSP_SERVER/GSP_SERVER/Object.h
@@ -5,13 +5,13 @@ class CObject
 {
 public:
 	char m_name[NAME_SIZE];
-	short m_id;
+	int m_id;
 	short m_x;
 	short m_y;
-	short m_exp;
-	short m_level;
-	short m_hp;
-	short m_maxhp;
+	int m_exp;
+	int m_level;
+	int m_hp;
+	int m_maxhp;
 
 	mutex m_s_lock;
 	atomic<S_STATE> m_state;
